Avoid configure failures
llvm builds many of its configure test programs with Werror and
Wunused-command-line-argument and the feliminate-unused-debug-types
cflag causes the latter to trigger in some cases (fpic being one)
making a number of compiler flag tests to fail.

Removing the feliminate-unused-debug-types flag makes the bootstrap
build succeed (and it is not used in later builds anyway).

Signed-off-by: William Douglas <william.douglas@intel.com>